// generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
	ID: { input: string; output: string }
	String: { input: string; output: string }
	Boolean: { input: boolean; output: boolean }
	Int: { input: number; output: number }
	Float: { input: number; output: number }
	DateTime: { input: any; output: any }
	ObjectId: { input: any; output: any }
}
// ...
export type Content = {
	__typename?: "Content"
	/** The HTML version of the content. */
	html: Scalars["String"]["output"]
	/** The Markdown version of the content. */
	markdown: Scalars["String"]["output"]
	/** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
	text: Scalars["String"]["output"]
}
// ...
export type DomainInfo = {
	__typename?: "DomainInfo"
	/** The domain of the publication. */
	domain?: Maybe<DomainStatus>
	/**
	 * The subdomain of the publication on hashnode.dev.
	 *
	 * It will redirect to you custom domain if it is present and ready.
	 */
	hashnodeSubdomain?: Maybe<Scalars["String"]["output"]>
	/** The www prefixed domain of the publication. Says if redirect to www domain is configured. */
	wwwPrefixedDomain?: Maybe<DomainStatus>
}
// ...
export type DraftConnection = Connection & {
	__typename?: "DraftConnection"
	/** A list of edges of drafts connection. */
	edges: Array<DraftEdge>
	/** Information to aid in pagination. */
	pageInfo: PageInfo
	/** The total number of documents in the connection. */
	totalDocuments: Scalars["Int"]["output"]
}
// ...
export type EmailImport = {
	__typename?: "EmailImport"
	/** Contains information about the current import example if it is in progress or has finished, date started, etc */
	currentImport?: Maybe<EmailCurrentImport>
}
// ...
export type Node = {
	/** The ID of the node. */
	id: Scalars["ID"]["output"]
}

/** Information to help in open graph related meta tags. */
export type OpenGraphMetaData = {
	__typename?: "OpenGraphMetaData"
	/** The image used in og:image tag for SEO purposes. */
	image?: Maybe<Scalars["String"]["output"]>
}
// ...
export type Publication = Node & {
	__typename?: "Publication"
	/** The about section of the publication. */
	about?: Maybe<Content>
	/** The author who owns the publication. */
	author: User
	/** The canonical URL of the publication. */
	canonicalURL: Scalars["String"]["output"]
	/** The description of the publication, used in og:description meta tag. */
	descriptionSEO?: Maybe<Scalars["String"]["output"]>
	/** The title of the publication. Shown in blog home page. */
	displayTitle?: Maybe<Scalars["String"]["output"]>
	/** Domain information of the publication. */
	domainInfo: DomainInfo
	/** Returns the list of drafts in the publication. */
	drafts: DraftConnection
	/** Returns the publication's email imports, used with newsletter feature. */
	emailImport?: Maybe<EmailImport>
	/** The favicon of the publication. Used in browser tab. */
	favicon?: Maybe<Scalars["String"]["output"]>
	/** Object containing information about beta features enabled for the publication. */
	features: PublicationFeatures
	/** Total number of followers of the publication. */
	followersCount?: Maybe<Scalars["Int"]["output"]>
	/** Whether the publication has earned any badges or not. */
	hasBadges: Scalars["Boolean"]["output"]
	/** Color code of the header color of the publication. Used to style header of blog. */
	headerColor?: Maybe<Scalars["String"]["output"]>
	/** The ID of the publication. */
	id: Scalars["ID"]["output"]
	/** Summary of the contact information and information related to copyrights, usually used in German-speaking countries. */
	imprint?: Maybe<Scalars["String"]["output"]>
	/** The integrations connected to the publication. */
	integrations?: Maybe<PublicationIntegrations>
	/** Returns true if GitHub backup is configured and active and false otherwise. */
	isGitHubBackupEnabled: Scalars["Boolean"]["output"]
	/** A flag to indicate if the publication is using Headless CMS. This can be used to check if the post redirect needs authentication. */
	isHeadless: Scalars["Boolean"]["output"]
	/** True if the publication is a team publication and false otherwise. */
	isTeam: Scalars["Boolean"]["output"]
	/** Links to the publication's social media profiles. */
	links?: Maybe<PublicationLinks>
	/** The meta tags associated with the publication. */
	metaTags?: Maybe<Scalars["String"]["output"]>
	/** Information about the publication's Open Graph metadata i.e. image. */
	ogMetaData: OpenGraphMetaData
	/** Returns the pinned post of the publication. */
	pinnedPost?: Maybe<Post>
	/** Returns the post with the given slug. */
	post?: Maybe<Post>
	/** Returns the list of posts in the publication. */
	posts: PublicationPostConnection
	/** The publication preferences around layout, theme and other personalisations. */
	preferences: Preferences
	/** Configured redirection rules for the publication. */
	redirectionRules: Array<RedirectionRule>
	/** Returns the scheduled drafts of the publication. */
	scheduledDrafts: DraftConnection
	/** Returns series by slug in the publication. */
	series?: Maybe<Series>
	/** Returns the list of series in the publication. */
	seriesList: SeriesConnection
	/** Contains the publication's sponsorships information. */
	sponsorship?: Maybe<PublicationSponsorship>
	/** Returns the static page with the given slug. */
	staticPage?: Maybe<StaticPage>
	/** Returns a list of static pages in the publication. */
	staticPages: StaticPageConnection
	/** Returns the list of submitted drafts in the publication. */
	submittedDrafts: DraftConnection
	/**
	 * The title of the publication.
	 * Title is used as logo if logo is not provided.
	 */
	title: Scalars["String"]["output"]
	/** The domain of the publication. Used to access publication. Example https://johndoe.com */
	url: Scalars["String"]["output"]
	/** Determines the structure of the post URLs. */
	urlPattern: UrlPattern
}
// ...
export type PublicationFeatures = {
	__typename?: "PublicationFeatures"
	/** Audio player for blog posts. */
	audioBlog: AudioBlogFeature
	/** Individual styling for the publication. */
	customCSS: CustomCssFeature
	/** Newsletter feature for the publication which adds a `/newsletter` route for collecting subscribers and allows sending out newsletters. */
	newsletter: NewsletterFeature
	/** Show the read time for blog posts. */
	readTime: ReadTimeFeature
	/** Widget that shows up if a text on a blog post is selected. Allows for easy sharing or copying of the selected text. */
	textSelectionSharer: TextSelectionSharerFeature
	/** Show the view count for blog posts. */
	viewCount: ViewCountFeature
}

/**
 * Contains the publication's integrations.
 * Used to connect the publication with third party services like Google Analytics, Facebook Pixel, etc.
 */
export type PublicationIntegrations = {
	__typename?: "PublicationIntegrations"
	/** Custom domain for integration with Fathom Analytics. */
	fathomCustomDomain?: Maybe<Scalars["String"]["output"]>
	/** A flag indicating if the custom domain is enabled for integration with Fathom Analytics. */
	fathomCustomDomainEnabled?: Maybe<Scalars["Boolean"]["output"]>
	/** Fathom Analytics Site ID for integration with Fathom Analytics. */
	fathomSiteID?: Maybe<Scalars["String"]["output"]>
	/** FB Pixel ID for integration with Facebook Pixel. */
	fbPixelID?: Maybe<Scalars["String"]["output"]>
	/** Google Analytics Tracking ID for integration with Google Analytics. */
	gaTrackingID?: Maybe<Scalars["String"]["output"]>
	/** Hotjar Site ID for integration with Hotjar. */
	hotjarSiteID?: Maybe<Scalars["String"]["output"]>
	/** Matomo Site ID for integration with Matomo Analytics. */
	matomoSiteID?: Maybe<Scalars["String"]["output"]>
	/** Matomo URL for integration with Matomo Analytics. */
	matomoURL?: Maybe<Scalars["String"]["output"]>
	/** A flag indicating if the custom domain is enabled for integration with Plausible Analytics. */
	plausibleAnalyticsEnabled?: Maybe<Scalars["Boolean"]["output"]>
	/** The ID for the Hashnode-provided Umami analytics instance. */
	umamiWebsiteUUID?: Maybe<Scalars["String"]["output"]>
	/** Web Monetization Payment Pointer for integration with Web Monetization. */
	wmPaymentPointer?: Maybe<Scalars["String"]["output"]>
}
// ...
export type PublicationLinks = {
	__typename?: "PublicationLinks"
	/** Daily.dev URL of the publication. */
	dailydev?: Maybe<Scalars["String"]["output"]>
	/** GitHub URL of the publication. */
	github?: Maybe<Scalars["String"]["output"]>
	/** Hashnode profile of author of the publication. */
	hashnode?: Maybe<Scalars["String"]["output"]>
	/** Instagram URL of the publication. */
	instagram?: Maybe<Scalars["String"]["output"]>
	/** LinkedIn URL of the publication. */
	linkedin?: Maybe<Scalars["String"]["output"]>
	/** Mastodon URL of the publication. */
	mastodon?: Maybe<Scalars["String"]["output"]>
	/** Twitter URL of the publication. */
	twitter?: Maybe<Scalars["String"]["output"]>
	/** Website URL of the publication. */
	website?: Maybe<Scalars["String"]["output"]>
	/** YouTube URL of the publication. */
	youtube?: Maybe<Scalars["String"]["output"]>
}
// ...
export type PublicationPostConnection = Connection & {
	__typename?: "PublicationPostConnection"
	/** A list of edges containing Post information */
	edges: Array<PostEdge>
	/** Information for pagination in Post connection. */
	pageInfo: PageInfo
	/** The total number of documents in the connection. */
	totalDocuments: Scalars["Int"]["output"]
}
// ...
export type User = IUser &
	Node & {
		__typename?: "User"
		/**
		 * Whether or not the user is an ambassador.
		 * @deprecated Ambassadors program no longer active
		 */
		ambassador: Scalars["Boolean"]["output"]
		/** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
		availableFor?: Maybe<Scalars["String"]["output"]>
		/** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
		badges: Array<Badge>
		/** The bio of the user. Visible in about me section of the user's profile. */
		bio?: Maybe<Content>
		/**
		 * The bio of the user. Visible in about me section of the user's profile.
		 * @deprecated Will be removed on 26/10/2023. Use bio instead of bioV2
		 */
		bioV2?: Maybe<Content>
		/** The date the user joined Hashnode. */
		dateJoined?: Maybe<Scalars["DateTime"]["output"]>
		/** Whether or not the user is deactivated. */
		deactivated: Scalars["Boolean"]["output"]
		/** The number of users that follow the requested user. Visible in the user's profile. */
		followersCount: Scalars["Int"]["output"]
		/**
		 * Wether or not the authenticated user follows this user.
		 * Returns false if the authenticated user this user.
		 */
		following: Scalars["Boolean"]["output"]
		/** The number of users that this user is following. Visible in the user's profile. */
		followingsCount: Scalars["Int"]["output"]
		/**
		 * Wether or not this user follows the authenticated user.
		 * Returns false if the authenticated user this user.
		 */
		followsBack: Scalars["Boolean"]["output"]
		/** The ID of the user. It can be used to identify the user. */
		id: Scalars["ID"]["output"]
		/** Wether or not this is a pro user. */
		isPro: Scalars["Boolean"]["output"]
		/** The location of the user. */
		location?: Maybe<Scalars["String"]["output"]>
		/** The name of the user. */
		name: Scalars["String"]["output"]
		/** The URL to the profile picture of the user. */
		profilePicture?: Maybe<Scalars["String"]["output"]>
		/** Publications associated with the user. Includes personal and team publications. */
		publications: UserPublicationsConnection
		/** The social media links of the user. Shown on the user's profile. */
		socialMediaLinks?: Maybe<SocialMediaLinks>
		/** The tagline of the user. Shown on the user's profile below the name. */
		tagline?: Maybe<Scalars["String"]["output"]>
		/** Returns a list of tags that the user follows. */
		tagsFollowing: Array<Tag>
		/** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
		username: Scalars["String"]["output"]
	}